requirements.txt
fastapi==0.88.0
uvicorn==0.18.3
SQLAlchemy==1.4.39
pydantic==1.9.1
boto3==1.24.28
opencv-python==4.5.5.64
tensorflow==2.8.0
numpy==1.22.4
scikit-learn==1.1.1
pymysql==1.0.2
aiofiles==0.8.0
matplotlib==3.5.2
platformdirs==2.5.2
python-dotenv==0.20.0


